+++ client/src/pages/databases/tree/index.tsx
@@ -2,36 +2,19 @@ import { useState, useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
 import { SimpleTreeView, TreeItem } from '@mui/x-tree-view';
 import icons from '@/components/icons/Icons';
-import {
-  Theme,
-  Tooltip,
-  Typography,
-  MenuItem,
-  Grow,
-  Popover,
-} from '@mui/material';
-import { useNavigate, Params } from 'react-router-dom';
+import { Tooltip, Typography, Grow, Popover } from '@mui/material';
+import { useNavigate } from 'react-router-dom';
 import { CollectionObject } from '@server/types';
 import clcx from 'clsx';
 import { formatNumber } from '@/utils';
-import { makeStyles } from '@mui/styles';
-
-export type TreeNodeType = 'db' | 'collection' | 'partition' | 'segment';
-
-export interface DatabaseTreeItem {
-  children?: DatabaseTreeItem[];
-  id: string;
-  name: string;
-  type: TreeNodeType;
-  expanded?: boolean;
-  data?: CollectionObject;
-}
-
-interface DatabaseToolProps {
-  database: string;
-  collections: CollectionObject[];
-  params: Readonly<Params<string>>;
-}
+import { useStyles } from './style';
+import {
+  DatabaseTreeItem,
+  TreeNodeType,
+  DatabaseToolProps,
+  ContextMenu,
+} from './types';
+import { TreeContextMenu } from './TreeContextMenu';
 
 // get expanded nodes from data
 const getExpanded = (nodes: DatabaseTreeItem[]) => {
@@ -46,108 +29,6 @@ const getExpanded = (nodes: DatabaseTreeItem[]) => {
   });
   return expanded;
 };
-
-const useStyles = makeStyles((theme: Theme) => ({
-  root: {
-    fontSize: '15px',
-    color: theme.palette.text.primary,
-    backgroundColor: theme.palette.background.default,
-    '& .MuiTreeItem-iconContainer': {
-      width: 'auto',
-    },
-    '& .MuiTreeItem-group': {
-      marginLeft: 0,
-      '& .MuiTreeItem-content': {
-        padding: '0 0 0 8px',
-      },
-    },
-    '& .MuiTreeItem-label:hover': {
-      backgroundColor: 'none',
-    },
-    '& .MuiTreeItem-content': {
-      width: 'auto',
-      padding: '0',
-      '&.Mui-focused': {
-        backgroundColor: 'rgba(10, 206, 130, 0.08)',
-      },
-      '&.Mui-selected': {
-        backgroundColor: 'rgba(10, 206, 130, 0.28)',
-      },
-      '&.Mui-focused.Mui-selected': {
-        backgroundColor: 'rgba(10, 206, 130, 0.28) !important',
-      },
-
-      '&:hover': {
-        backgroundColor: 'rgba(10, 206, 130, 0.08)',
-      },
-      '& .MuiTreeItem-label': {
-        background: 'none',
-      },
-    },
-  },
-  treeItem: {
-    '& .MuiTreeItem-iconContainer': {
-      color: '#888',
-    },
-    '& .right-selected-on': {
-      '& .MuiTreeItem-content': {
-        backgroundColor: 'rgba(10, 206, 130, 0.08)',
-        '&.Mui-selected': {
-          backgroundColor: 'rgba(10, 206, 130, 0.28) !important',
-        },
-      },
-    },
-  },
-  collectionNode: {
-    minHeight: '24px',
-    lineHeight: '24px',
-    position: 'relative',
-  },
-  collectionName: {
-    display: 'flex',
-    alignItems: 'center',
-    width: 'calc(100% - 45px)',
-    '& .collectionName': {
-      minWidth: 36,
-    },
-  },
-  dbName: {
-    width: 'calc(100% - 30px)',
-  },
-  count: {
-    fontSize: '13px',
-    fontWeight: 500,
-    marginLeft: theme.spacing(0.5),
-    color: theme.palette.text.secondary,
-    pointerEvents: 'none',
-  },
-  dot: {
-    width: '8px',
-    height: '8px',
-    borderRadius: '50%',
-    position: 'absolute',
-    left: 160,
-    top: 8,
-    zIndex: 1,
-    pointerEvents: 'none',
-  },
-  loaded: {
-    border: `1px solid ${theme.palette.primary.main}`,
-    backgroundColor: theme.palette.primary.main,
-  },
-  unloaded: {
-    border: `1px solid ${theme.palette.primary.main}`,
-    background: '#fff !important',
-  },
-  loading: {
-    border: `1px solid ${theme.palette.primary.light}`,
-    backgroundColor: `${theme.palette.primary.light} !important`,
-  },
-  noIndex: {
-    border: `1px solid ${theme.palette.text.disabled}`,
-    backgroundColor: theme.palette.text.disabled,
-  },
-}));
 
 const CollectionNode: React.FC<{ data: CollectionObject }> = ({ data }) => {
   // i18n collectionTrans
@@ -194,11 +75,7 @@ const CollectionNode: React.FC<{ data: CollectionObject }> = ({ data }) => {
 
 const DatabaseTree: React.FC<DatabaseToolProps> = props => {
   // state
-  const [contextMenu, setContextMenu] = useState<{
-    mouseX: number;
-    mouseY: number;
-    nodeId: string | null;
-  } | null>(null);
+  const [contextMenu, setContextMenu] = useState<ContextMenu | null>(null);
   // props
   const { database, collections, params } = props;
 
@@ -242,25 +119,21 @@ const DatabaseTree: React.FC<DatabaseToolProps> = props => {
     setContextMenu(null);
   };
 
-  const handleContextMenu = (event: any, nodeId: string) => {
+  const handleContextMenu = (event: any, nodeId: string, nodeType: string) => {
+    // prevent default
     event.preventDefault();
     event.stopPropagation();
+
     setContextMenu({
       mouseX: event.clientX - 2,
       mouseY: event.clientY - 4,
       nodeId,
+      nodeType: nodeType as TreeNodeType,
     });
   };
 
   const handleClose = () => {
     setContextMenu(null);
-  };
-
-  const handleMenuClick = (action: string) => {
-    if (contextMenu?.nodeId) {
-      console.log(`Action ${action} on node ${contextMenu.nodeId}`);
-    }
-    handleClose();
   };
 
   // render children
@@ -298,7 +171,7 @@ const DatabaseTree: React.FC<DatabaseToolProps> = props => {
             icon: CollectionIcon,
           }}
           label={<CollectionNode data={node.data!} />}
-          onContextMenu={event => handleContextMenu(event, node.id)}
+          onContextMenu={event => handleContextMenu(event, node.id, node.type)}
           className={clcx(classes.treeItem, {
             ['right-selected-on']: contextMenu?.nodeId === node.id,
           })}
@@ -357,7 +230,9 @@ const DatabaseTree: React.FC<DatabaseToolProps> = props => {
                 onNodeClick(tree);
               }
             }}
-            onContextMenu={event => handleContextMenu(event, tree.id)}
+            onContextMenu={event =>
+              handleContextMenu(event, tree.id, tree.type)
+            }
           >
             {tree.children && tree.children.length > 0
               ? renderTree(tree.children)
@@ -376,18 +251,12 @@ const DatabaseTree: React.FC<DatabaseToolProps> = props => {
         }
         TransitionComponent={Grow}
         transitionDuration={0}
-        // disableAutoFocusItem
-        sx={{ pointerEvents: 'none' }} // 使 Popover 外部不可交互
+        sx={{ pointerEvents: 'none' }}
         PaperProps={{
-          sx: { pointerEvents: 'auto', boxShadow: 3 },
+          sx: { pointerEvents: 'auto', boxShadow: 4, borderRadius: 2 },
         }}
       >
-        <MenuItem onClick={() => handleMenuClick('Rename')}>
-          Rename Collection
-        </MenuItem>
-        <MenuItem onClick={() => handleMenuClick('Delete')}>
-          Delete Collection
-        </MenuItem>
+        <TreeContextMenu onClick={handleClose} contextMenu={contextMenu!} />
       </Popover>
     </>
   );
