+++ client/src/pages/databases/tree/types.ts
@@ -1,0 +1,26 @@
+import { CollectionObject } from '@server/types';
+import { Params } from 'react-router-dom';
+
+export type TreeNodeType = 'db' | 'collection' | 'partition' | 'segment';
+
+export interface DatabaseTreeItem {
+  children?: DatabaseTreeItem[];
+  id: string;
+  name: string;
+  type: TreeNodeType;
+  expanded?: boolean;
+  data?: CollectionObject;
+}
+
+export interface DatabaseToolProps {
+  database: string;
+  collections: CollectionObject[];
+  params: Readonly<Params<string>>;
+}
+
+export type ContextMenu = {
+  mouseX: number; // x position
+  mouseY: number; // y position
+  nodeId: string | null; // node id
+  nodeType: TreeNodeType; // node type
+};
